pycdel

+++ project/package/pycdel.py
@@ -1,31 +1,37 @@
 import argparse
+import builtins
 import os
 
 _PYCACHE_DIR_NAME = '__pycache__'
 _PYCACHE_FILE_NAME_EXT = '.pyc'
+_CONFIRM_REMOVE_FILE_DEFAULT = lambda fn: (input(f'Do you want to remove file {repr(fn)}? [y/n] ') == 'y')
 
-def del_cache_dir(d:str):
+def del_cache_dir(d:str,
+                  confirm=_CONFIRM_REMOVE_FILE_DEFAULT,
+                  debug=False):
 
-    print(f'Directory: {repr(d)}')
+    print = (lambda *a,**ka: None) if not debug else builtins.print
+    print(f'Directory: {d}')
     to_delete_dir = True
     for fn in os.listdir(d):
 
         if fn.endswith(_PYCACHE_FILE_NAME_EXT):
 
             fn_full = os.path.join(d, fn)
-            os.remove(fn_full)
+            if confirm(fn_full):
 
-        else:
+                print(f'   Remove file {fn_full}')
+                os.remove(fn_full)
 
-            to_delete_dir = False
+    if not os.listdir(d):
 
-    if to_delete_dir:
-
+        print(f'Remove empty directory {d}')
         os.rmdir(d)
-        print(f'Remove {repr(d)}')
 
 def do_it(wd       :str,
-          recursive:bool=False):
+          recursive    =False,
+          confirm      =_CONFIRM_REMOVE_FILE_DEFAULT,
+          debug        =False):
     
     if not recursive:
 
@@ -38,7 +44,7 @@ def do_it(wd       :str,
 
             if dn == _PYCACHE_DIR_NAME:
 
-                del_cache_dir(os.path.join(dp, dn))
+                del_cache_dir(os.path.join(dp, dn), confirm=confirm, debug=debug)
 
 def main():
 
@@ -47,16 +53,30 @@ def main():
 
         WORKING_DIR = 'wd'
         RECURSIVE   = 'r'
+        NO_ASKING   = 'noask'
+        DEBUG       = 'debug'
 
-    ap.add_argument(f'--{A.WORKING_DIR}')
+    ap.add_argument(f'--{A.WORKING_DIR}',
+                    help='working directory, from which to search and remove cache')
     ap.add_argument(f'--{A.RECURSIVE}',
+                    help='find and remove cache recursively i.e. in sub-directories',
+                    action='store_true')
+    ap.add_argument(f'--{A.NO_ASKING}',
+                    help='do not ask before deleting\nMake sure that you do want to delete ALL cache files, before setting this options.',
+                    action='store_true')
+    ap.add_argument(f'--{A.DEBUG}',
+                    help='debug i.e. print more information about what is happening',
                     action='store_true')
     # read args
     get = ap.parse_args().__getattribute__
-    wd = get(A.WORKING_DIR)
-    r  = get(A.RECURSIVE)
+    wd     = get(A.WORKING_DIR)
+    r      = get(A.RECURSIVE)
+    no_ask = get(A.NO_ASKING)
+    debug  = get(A.DEBUG)
     # do it
     do_it(wd       =wd if wd is not None else os.getcwd(),
-          recursive=r)
+          recursive=r,
+          confirm  =_CONFIRM_REMOVE_FILE_DEFAULT if not no_ask else (lambda fn: True),
+          debug    =debug)
 
 if __name__ == '__main__': main()
